requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
openai>=1.12.0
python-dotenv
requests 